Add --xmp-metadata option
